Implements start container.
This PR makes a workaround to start container by restarting the pod.
